python_ai_service/requirements.txt
flask
flask-cors
pymongo
pandas
openai
python-dotenv
certifi
google-generativeai
